requirements.txt
gradio
langchain-community
langchain
ollama
pymupdf4llm
chromadb
langchain-openai
langchain-community
faiss-cpu

